+++ ExpandWorld/compatibility/General.cs
@@ -6,7 +6,7 @@ using HarmonyLib;
 namespace ExpandWorld;
 
 [HarmonyPatch]
-public class TryParse
+public class TryParseBiome
 {
   static IEnumerable<MethodBase> TargetMethods()
   {
@@ -22,14 +22,38 @@ public class TryParse
   }
 }
 
+[HarmonyPatch]
+public class TryParseTheme
+{
+  static IEnumerable<MethodBase> TargetMethods()
+  {
+    return typeof(Enum).GetMethods()
+        .Where(method => method.Name.StartsWith("TryParse"))
+        .Select(method => method.MakeGenericMethod(typeof(Room.Theme)))
+        .Cast<MethodBase>();
+  }
+  static bool Prefix(string value, ref Room.Theme result, ref bool __result)
+  {
+    __result = RoomLoading.TryGetTheme(value, out result);
+    return false;
+  }
+}
 [HarmonyPatch(typeof(Enum), nameof(Enum.GetValues))]
 public class GetValues
 {
   static bool Prefix(Type enumType, ref Array __result)
   {
-    if (enumType != typeof(Heightmap.Biome)) return true;
-    __result = BiomeManager.BiomeToDisplayName.Keys.ToArray();
-    return false;
+    if (enumType == typeof(Heightmap.Biome))
+    {
+      __result = BiomeManager.BiomeToDisplayName.Keys.ToArray();
+      return false;
+    }
+    if (enumType == typeof(Room.Theme))
+    {
+      __result = RoomLoading.ThemeToName.Keys.ToArray();
+      return false;
+    }
+    return true;
   }
 }
 [HarmonyPatch(typeof(Enum), nameof(Enum.GetNames))]
@@ -37,9 +61,17 @@ public class GetNames
 {
   static bool Prefix(Type enumType, ref string[] __result)
   {
-    if (enumType != typeof(Heightmap.Biome)) return true;
-    __result = BiomeManager.BiomeToDisplayName.Values.ToArray();
-    return false;
+    if (enumType == typeof(Heightmap.Biome))
+    {
+      __result = BiomeManager.BiomeToDisplayName.Values.ToArray();
+      return false;
+    }
+    if (enumType == typeof(Room.Theme))
+    {
+      __result = RoomLoading.ThemeToName.Values.ToArray();
+      return false;
+    }
+    return true;
   }
 }
 [HarmonyPatch(typeof(Enum), nameof(Enum.GetName))]
@@ -47,12 +79,23 @@ public class GetName
 {
   static bool Prefix(Type enumType, object value, ref string __result)
   {
-    if (enumType != typeof(Heightmap.Biome)) return true;
-    if (BiomeManager.TryGetDisplayName((Heightmap.Biome)value, out var result))
-      __result = result;
-    else
-      __result = "None";
-    return false;
+    if (enumType == typeof(Heightmap.Biome))
+    {
+      if (BiomeManager.TryGetDisplayName((Heightmap.Biome)value, out var result))
+        __result = result;
+      else
+        __result = "None";
+      return false;
+    }
+    if (enumType == typeof(Room.Theme))
+    {
+      if (RoomLoading.ThemeToName.TryGetValue((Room.Theme)value, out var result))
+        __result = result;
+      else
+        __result = ((int)value).ToString();
+      return false;
+    }
+    return true;
   }
 }
 [HarmonyPatch(typeof(Enum), nameof(Enum.Parse), typeof(Type), typeof(string))]
@@ -60,12 +103,25 @@ public class EnumParse
 {
   static bool Prefix(Type enumType, string value, ref object __result)
   {
-    if (enumType != typeof(Heightmap.Biome)) return true;
-    if (BiomeManager.TryGetBiome(value, out var biome))
-      __result = biome;
-    else
-      return true; // Let the original function handle the throwing.
-    return false;
+    if (enumType == typeof(Heightmap.Biome))
+    {
+      if (BiomeManager.TryGetBiome(value, out var biome))
+      {
+        __result = biome;
+        return false;
+      }
+      // Let the original function handle the throwing.
+    }
+    if (enumType == typeof(Room.Theme))
+    {
+      if (RoomLoading.TryGetTheme(value, out var theme))
+      {
+        __result = theme;
+        return false;
+      }
+      // Let the original function handle the throwing.
+    }
+    return true;
   }
 }
 [HarmonyPatch(typeof(Enum), nameof(Enum.Parse), typeof(Type), typeof(string), typeof(bool))]
@@ -73,11 +129,24 @@ public class ParseIgnoreCase
 {
   static bool Prefix(Type enumType, string value, ref object __result)
   {
-    if (enumType != typeof(Heightmap.Biome)) return true;
-    if (BiomeManager.TryGetBiome(value, out var biome))
-      __result = biome;
-    else
-      return true; // Let the original function handle the throwing.
-    return false;
+    if (enumType == typeof(Heightmap.Biome))
+    {
+      if (BiomeManager.TryGetBiome(value, out var biome))
+      {
+        __result = biome;
+        return false;
+      }
+      // Let the original function handle the throwing.
+    }
+    if (enumType == typeof(Room.Theme))
+    {
+      if (RoomLoading.TryGetTheme(value, out var theme))
+      {
+        __result = theme;
+        return false;
+      }
+      // Let the original function handle the throwing.
+    }
+    return true;
   }
 }